[MERGE] DEV

+++ controller/chatGPT/deleteContent.js
@@ -1,4 +1,4 @@
-// PUT /chatgpt/list/:list_id/content/:content_id
+// PUT /chatgpt/list/:list_id/content
 // 대화 내용 삭제
 import models from '../../models/index.js';
 import { Op } from 'sequelize';
@@ -6,21 +6,31 @@ import { Op } from 'sequelize';
 const deleteContent = async (req, res, next) => {
   try {
     const listId = Number(req.params.list_id);
-    const contentId = Number(req.params.content_id);
+    const { contentIdList } = req.body;
+    // const contentId = Number(req.params.content_id);
 
-    const Content = await models.ChatGPTContent.findOne({
-      where: {
-        id: contentId
-      }
-    });
+    if (!contentIdList) {
+      return res.status(400).send('EMPTY_CONTENT_ID_LIST');
+    }
 
-    const questionNum = Content.questionNum;
-    console.log('questionNum = ', questionNum);
-    await models.ChatGPTContent.destroy({
-      where: {
-        [Op.and]: [{ listId: listId }, { questionNum: questionNum }]
+    for (let i = 0; i < contentIdList.length; i++) {
+      const Content = await models.ChatGPTContent.findOne({
+        where: {
+          id: contentIdList[i]
+        }
+      });
+
+      const questionNum = Content.questionNum;
+      if (questionNum === 0) {
+        return res.status(400).send('UNABLE_TO_DELETE_CONTENT');
+      } else {
+        await models.ChatGPTContent.destroy({
+          where: {
+            [Op.and]: [{ listId: listId }, { questionNum: questionNum }]
+          }
+        });
       }
-    });
+    }
 
     return res.status(200).send('DELETE_CONTENT_SUCCESS');
   } catch (err) {
